+++ src/sdc11073/consumer/manipulator.py
@@ -1,10 +1,11 @@
+"""A manipulator to manipulate outgoing messages."""
+
 from __future__ import annotations
 
 from typing import TYPE_CHECKING, Protocol
 
 if TYPE_CHECKING:
     from sdc11073 import xml_utils
-
     from sdc11073.pysoap.soapenvelope import Soap12Envelope
 
 
@@ -23,5 +24,5 @@ class RequestManipulatorProtocol(Protocol):
     def manipulate_domtree(self, domtree: xml_utils.LxmlElement) -> xml_utils.LxmlElement | None:
         """Manipulate on etree.Element level."""
 
-    def manipulate_string(self, xml_string: str) -> str | None:
+    def manipulate_string(self, xml_string: bytes) -> bytes | None:
         """Manipulate on string level."""