+++ src/sdc11073/pysoap/msgfactory.py
@@ -1,44 +1,61 @@
+"""The msgfactory module contains the MessageFactory that creates soap messages."""
+
 from __future__ import annotations
 
 from io import BytesIO
-from typing import Optional, Union, List, Type, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
 from lxml import etree
 
-from .msgreader import validate_node
-from .soapenvelope import Soap12Envelope
+from sdc11073.pysoap.msgreader import validate_node
+from sdc11073.pysoap.soapenvelope import Soap12Envelope
 from sdc11073.schema_resolver import mk_schema_validator
 
 if TYPE_CHECKING:
+    from collections.abc import Iterable
+
+    from sdc11073 import xml_utils
+    from sdc11073.consumer.manipulator import RequestManipulatorProtocol
+    from sdc11073.definitions_base import BaseDefinitions
+    from sdc11073.dispatch.request import RequestData
+    from sdc11073.loghelper import LoggerAdapter
+    from sdc11073.namespaces import PrefixNamespace
     from sdc11073.xml_types.addressing_types import HeaderInformationBlock
     from sdc11073.xml_types.msg_types import MessageType
-    from sdc11073.definitions_base import BaseDefinitions
-    from sdc11073.namespaces import PrefixNamespace
-    from sdc11073 import xml_utils
 
 
 class CreatedMessage:
-    def __init__(self, message: Soap12Envelope, msg_factory):
+    """Represents a soap message created by a MessageFactory together with the factory that created it."""
+
+    def __init__(self, message: Soap12Envelope, msg_factory: MessageFactory):
         self.p_msg = message
         self.msg_factory = msg_factory
 
-    def serialize(self, pretty=False, request_manipulator=None, validate=True):
+    def serialize(
+        self,
+        pretty: bool = False,
+        request_manipulator: RequestManipulatorProtocol | None = None,
+        validate: bool = True,
+    ) -> bytes:
+        """Serialize the message to its raw XML representation."""
         return self.msg_factory.serialize_message(self, pretty, request_manipulator, validate)
 
 
-# pylint: disable=no-self-use
-
-
 class MessageFactory:
-    """This class creates soap messages. It is used in two phases:
-     1) call one of the mk_xxx methods. All return a CreatedMessage instance that contains the data provided in the call
-     2) call the serialize method of the CreatedMessage instance to get the xml representation
-     """
+    """Create soap messages.
 
-    def __init__(self, sdc_definitions: Type[BaseDefinitions],
-                 additional_schema_specs: Union[List[PrefixNamespace], None],
-                 logger,
-                 validate=True):
+    It is used in two phases:
+    1) call one of the mk_xxx methods. All return a CreatedMessage instance that contains the data provided in the call
+    2) call the serialize method of the CreatedMessage instance to get the xml representation
+    """
+
+    def __init__(
+        self,
+        sdc_definitions: type[BaseDefinitions],
+        additional_schema_specs: Iterable[PrefixNamespace] | None,
+        logger: LoggerAdapter,
+        validate: bool = True,
+    ):
         self.schema_specs = [entry.value for entry in sdc_definitions.data_model.ns_helper.prefix_enum]
         if additional_schema_specs is not None:
             self.schema_specs.extend(additional_schema_specs)
@@ -47,9 +64,14 @@ class MessageFactory:
         self._validate = validate
         self._xml_schema: etree.XMLSchema = mk_schema_validator(self.schema_specs, self.ns_hlp)
 
-    def serialize_message(self, message: CreatedMessage, pretty=False,
-                          request_manipulator=None, validate=True) -> bytes:
-        """
+    def serialize_message(
+        self,
+        message: CreatedMessage,
+        pretty: bool = False,
+        request_manipulator: RequestManipulatorProtocol | None = None,
+        validate: bool = True,
+    ) -> bytes:
+        """Serialize a CreatedMessage instance to bytes.
 
         :param message: a CreatedMessage instance
         :param pretty:
@@ -83,16 +105,16 @@ class MessageFactory:
         doc.write(tmp, encoding='UTF-8', xml_declaration=True, pretty_print=pretty)
         return tmp.getvalue()
 
-    def mk_soap_message(self,
-                        header_info: HeaderInformationBlock,
-                        payload: MessageType,
-                        ns_list: Optional[list] = None,
-                        use_defaults=True) -> CreatedMessage:
+    def mk_soap_message(
+        self,
+        header_info: HeaderInformationBlock,
+        payload: MessageType,
+        ns_list: Iterable[PrefixNamespace] | None = None,
+        use_defaults: bool = True,
+    ) -> CreatedMessage:
+        """Create a soap message for the given payload."""
         nsh = self.ns_hlp
-        if use_defaults:
-            ns_set = {nsh.S12, nsh.WSA, nsh.MSG, nsh.PM}  # default
-        else:
-            ns_set = set()
+        ns_set = {nsh.S12, nsh.WSA, nsh.MSG, nsh.PM} if use_defaults else set()  # default
         ns_set.update(payload.additional_namespaces)
         if ns_list:
             ns_set.update(ns_list)
@@ -102,9 +124,10 @@ class MessageFactory:
         soap_envelope.payload_element = payload.as_etree_node(payload.NODETYPE, my_ns_map)
         return CreatedMessage(soap_envelope, self)
 
-    def mk_soap_message_etree_payload(self,
-                                      header_info: HeaderInformationBlock,
-                                      payload_element: xml_utils.LxmlElement | None = None) -> CreatedMessage:
+    def mk_soap_message_etree_payload(
+        self, header_info: HeaderInformationBlock, payload_element: xml_utils.LxmlElement | None = None
+    ) -> CreatedMessage:
+        """Create a soap message with a raw etree payload element."""
         nsh = self.ns_hlp
         my_ns_map = nsh.partial_map(nsh.S12, nsh.WSE, nsh.WSA)
         soap_envelope = Soap12Envelope(my_ns_map)
@@ -112,10 +135,10 @@ class MessageFactory:
         soap_envelope.payload_element = payload_element
         return CreatedMessage(soap_envelope, self)
 
-    def mk_reply_soap_message(self,
-                              request,
-                              response_payload: MessageType,
-                              ns_map: Optional[list] = None) -> CreatedMessage:
+    def mk_reply_soap_message(
+        self, request: RequestData, response_payload: MessageType, ns_map: Iterable[PrefixNamespace] | None = None
+    ) -> CreatedMessage:
+        """Create a soap reply message for the given request."""
         nsh = self.ns_hlp
         ns_set = {nsh.S12, nsh.WSA, nsh.MSG, nsh.PM}  # default
         ns_set.update(response_payload.additional_namespaces)
@@ -124,11 +147,12 @@ class MessageFactory:
         my_ns_map = nsh.partial_map(*ns_set)
         soap_envelope = Soap12Envelope(my_ns_map)
         reply_address = request.message_data.p_msg.header_info_block.mk_reply_header_block(
-            action=response_payload.action)
+            action=response_payload.action
+        )
         soap_envelope.set_header_info_block(reply_address)
         soap_envelope.payload_element = response_payload.as_etree_node(response_payload.NODETYPE, my_ns_map)
         return CreatedMessage(soap_envelope, self)
 
-    def _validate_node(self, node):
+    def _validate_node(self, node: xml_utils.LxmlElement):
         if self._validate:
             validate_node(node, self._xml_schema, self._logger)